+++ tests/assert_point.rs
@@ -35,8 +35,8 @@ fn assert_equal_point() {
 
     impl Circuit for TestCircuit {
         fn circuit(&self, composer: &mut Composer) -> Result<(), Error> {
-            let w_p1 = composer.append_point(self.p1);
-            let w_p2 = composer.append_point(self.p2);
+            let w_p1 = composer.append_point(self.p1)?;
+            let w_p2 = composer.append_point(self.p2)?;
             composer.assert_equal_point(w_p1, w_p2);
 
             Ok(())
@@ -124,8 +124,8 @@ fn assert_equal_public_point() {
 
     impl Circuit for TestCircuit {
         fn circuit(&self, composer: &mut Composer) -> Result<(), Error> {
-            let w_point = composer.append_point(self.point);
-            composer.assert_equal_public_point(w_point, self.public);
+            let w_point = composer.append_point(self.point)?;
+            composer.assert_equal_public_point(w_point, self.public)?;
 
             Ok(())
         }