+++ tests/composer.rs
@@ -42,7 +42,7 @@ fn circuit_with_all_gates() {
             let w_b = composer.append_witness(self.b);
             let w_x = composer.append_witness(self.x);
             let w_y = composer.append_witness(self.y);
-            let w_z = composer.append_point(self.z);
+            let w_z = composer.append_point(self.z)?;
 
             let s = Constraint::new().mult(1).a(w_a).b(w_b);
 
@@ -50,13 +50,13 @@ fn circuit_with_all_gates() {
 
             composer.append_constant(15);
             composer.append_constant_point(self.z)?;
-            composer.append_public_point(self.z);
+            composer.append_public_point(self.z)?;
             composer.append_public(self.y);
 
             composer.assert_equal(w_x, r_w);
             composer.assert_equal_constant(w_x, 0, Some(self.x));
             composer.assert_equal_point(w_z, w_z);
-            composer.assert_equal_public_point(w_z, self.z);
+            composer.assert_equal_public_point(w_z, self.z)?;
 
             composer.gate_add(Constraint::new().left(1).right(1).a(w_a).b(w_b));